code refactor #1
