+++ src/components/BookList/BookDetail.js
@@ -1,0 +1,4 @@
+
+export default function BookDetail({ book }) {
+
+}
