+++ src/components/BookList/BookList.js
@@ -1,0 +1,45 @@
+import BookShow from '../BookShow/BookShow';
+import { useState, useEffect } from 'react';
+
+export default function BookList({ books }) {
+
+
+    // Infinite Scroll
+    const itemsPerPage = 9;
+
+    const [currentPage, setCurrentPage] = useState(1);
+    const [currentBooks, setCurrentBooks] = useState([]);
+
+    useEffect(() => {
+        setCurrentBooks(books.slice(0, currentPage * itemsPerPage));
+    }, [books, currentPage]);
+
+    useEffect(() => {
+    const observer = new IntersectionObserver(
+        (entries) => {
+        if (entries[0].isIntersecting) {
+            setCurrentPage((prev) => prev + 1);
+        }
+        },
+        { threshold: 1 }
+    );
+    observer.observe(document.querySelector("#end-of-list"));
+    }, []);
+
+    // Books mapping
+    const renderedBooks = currentBooks.map((book) => {
+        return(
+            <BookShow book={book} key={book.isbn13}/>
+    )})
+
+
+    // Modal
+
+    return(
+        <div className="background-color full-height px-5 pb-5">
+            <div className="d-flex flex-row flex-wrap">
+                {renderedBooks}  
+                <div id="end-of-list" />
+            </div>
+        </div>
+    )};
