+++ src/components/styles/card-style.css
@@ -3,6 +3,7 @@
     width: 30rem !important;
     box-shadow: 2px 4px 8px 0px rgb(0, 0, 0);
     margin: 3rem auto auto auto;
+    transition-duration: 1s;
 }
 
 .card-radius {
@@ -19,8 +20,11 @@
 }
 
 .card-img {
-    height: 247px !important;
+    height: 250px !important;
     width: auto !important;
     max-width: 169px;
     margin: 0 !important;
+    border-radius: 24px 0px 0px 24px!important;
+    box-shadow: 2px 4px 8px 0px rgb(0, 0, 0);
 }
+
